Refactor `admin.extensions.v1/components/application`
