+++ features/admin.roles.v2/models/policies.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2023, WSO2 LLC. (https://www.wso2.com). All Rights Reserved.
+ * Copyright (c) 2023-2024, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -16,4 +16,16 @@
  * under the License.
  */
 
-export * from "../../../../admin.application-roles.v1/models/application-roles";
+/**
+ * Policies enum.
+ */
+export enum Policy {
+    /**
+     * Role based policy.
+     */
+    ROLE = "RBAC",
+    /**
+     * No Policy.
+     */
+    NO = "NONE"
+}
